refactored, updated nunit libs
